+++ crates/matrix-sdk-crypto/src/store/integration_tests.rs
@@ -1603,26 +1603,26 @@ macro_rules! cryptostore_integration_tests_time {
                 assert!(acquired5.is_none());
 
                 // That's a nice test we got here, go take a little nap.
-                tokio::time::sleep(Duration::from_millis(50)).await;
+                matrix_sdk_common::sleep::sleep(Duration::from_millis(50)).await;
 
                 // Still too early.
                 let acquired55 = store.try_take_leased_lock(300, "key", "bob").await.unwrap();
                 assert!(acquired55.is_none()); // not acquired
 
                 // Ok you can take another nap then.
-                tokio::time::sleep(Duration::from_millis(250)).await;
+                matrix_sdk_common::sleep::sleep(Duration::from_millis(250)).await;
 
                 // At some point, we do get the lock.
                 let acquired6 = store.try_take_leased_lock(0, "key", "bob").await.unwrap();
                 assert_eq!(acquired6, Some(2)); // new lock generation!
 
-                tokio::time::sleep(Duration::from_millis(1)).await;
+                matrix_sdk_common::sleep::sleep(Duration::from_millis(1)).await;
 
                 // The other gets it almost immediately too.
                 let acquired7 = store.try_take_leased_lock(0, "key", "alice").await.unwrap();
                 assert_eq!(acquired7, Some(3)); // new lock generation!
 
-                tokio::time::sleep(Duration::from_millis(1)).await;
+                matrix_sdk_common::sleep::sleep(Duration::from_millis(1)).await;
 
                 // But when we take a longer lease…
                 let acquired8 = store.try_take_leased_lock(300, "key", "bob").await.unwrap();